Read a text file and print it to POSIX

+++ 0x15-file_io/0-read_textfile.c
@@ -32,5 +32,5 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	free(buffer);
 	close(o);
 
-	return (w);
+	return ((size_t)w);
 }
